Restructure to final filestructure.
