--- new/Hardware/NanoVNA_F_V2.py
@@ -1,16 +1,0 @@
-import logging
-
-from .NanoVNA import NanoVNA
-from .Serial import Interface
-
-logger = logging.getLogger(__name__)
-
-
-class NanoVNA_F_V2(NanoVNA):
-    name = "NanoVNA-F_V2"
-    screenwidth = 800
-    screenheight = 480
-
-    def __init__(self, iface: Interface):
-        super().__init__(iface)
-        self.sweep_max_freq_Hz = 3e9
