--- src/pynanovna/__init__.py
@@ -1,4 +1,0 @@
-from .pynanovna import *
-from .vis import *
-
-#  Needed to import the directory as a regular package.
