+++ src/pynanovna/hardware/AVNA.py
@@ -1,8 +1,12 @@
+import logging
+
 from .Serial import Interface
-from .VNA import VNA
+from .VNABase import VNABase
+
+logger = logging.getLogger(__name__)
 
 
-class AVNA(VNA):
+class AVNA(VNABase):
     name = "AVNA"
 
     def __init__(self, iface: Interface):
@@ -10,12 +14,12 @@ class AVNA(VNA):
         self.sweep_max_freq_Hz = 40e3
         self.features.add("Customizable data points")
 
-    def isValid(self):
+    def is_valid(self):
         return True
 
-    def resetSweep(self, start: int, stop: int):
+    def reset_sweep(self, start: int, stop: int):
         list(self.exec_command(f"sweep {start} {stop} {self.datapoints}"))
         list(self.exec_command("resume"))
 
-    def setSweep(self, start, stop):
+    def set_sweep(self, start, stop):
         list(self.exec_command(f"sweep {start} {stop} {self.datapoints}"))
