+++ src/pynanovna/hardware/JNCRadio_VNA_3G.py
@@ -1,5 +1,9 @@
+import logging
+
 from .NanoVNA import NanoVNA
 from .Serial import Interface
+
+logger = logging.getLogger(__name__)
 
 
 class JNCRadio_VNA_3G(NanoVNA):
@@ -14,7 +18,7 @@ class JNCRadio_VNA_3G(NanoVNA):
         super().__init__(iface)
         self.sweep_max_freq_Hz = 3e9
 
-    def setSweep(self, start, stop):
+    def set_sweep(self, start, stop):
         self.start = start
         self.stop = stop
         list(self.exec_command(f"scan {start} {stop} {self.datapoints}"))
