+++ src/pynanovna/hardware/NanoVNA_H4.py
@@ -1,5 +1,9 @@
+import logging
+
 from .Serial import Interface
 from .NanoVNA_H import NanoVNA_H
+
+logger = logging.getLogger(__name__)
 
 
 class NanoVNA_H4(NanoVNA_H):
@@ -14,3 +18,11 @@ class NanoVNA_H4(NanoVNA_H):
         self.sweep_method = "scan"
         if "Scan mask command" in self.features:
             self.sweep_method = "scan_mask"
+
+    # def read_features(self):
+    #     logger.debug("read_features")
+    #     super().read_features()
+    #     if self.readFirmware().find("DiSlord") > 0:
+    #         self.features.add("Customizable data points")
+    #         logger.info("VNA has 201 datapoints capability")
+    #         self.valid_datapoints = (201, 11, 51,101)
